requirement.txt
confluent-kafka
requests
beautifulsoup4
pandas
python-dotenv
psycopg2-binary
dbt-postgres

